Share formatDate function
